fix(merge): merge base_version instead of add
Adding base_version to the database violated the unique constraint on the
baseid column of that table. But since a user can provide a baseid as
part of the create step, this constraint isn't as important outside of
baseid being the primary key of base_version. So it is safe to merge.