Implement DRS v0.5.0 (closes #146)
This issues is being tracked in different locations. For Santa Cruz,
this commit closes [TLP-387](https://ucsc-cgl.atlassian.net/browse/TLP-387).
For Chicago, this closes PXD-1134.

This commit updates the behavior of DRS endpoints such that it
reflects the changes and features introduced in v0.5.0 (from v0.2.0).
Specifically, the following changes have been made:
* Functionality of `POST /dataobjects/list` has been moved to
  `GET /dataobjects`, and the parameters provided to this endpoint
  are now specified as query parameters instead of in the request body.
* Specification of required and optional parameters has been properly
  serialized in the schema.
* `checksum` parameter is now optional for ListDataBundles and
  ListDataObjects.
* Tests have also been updated to this effect.
* Minor changes for PEP8 compliance.

It is worth noting in this message that DOS is in the process of being
renamed to DRS (Data Repository Service) - while the decision has been
made, the process of percolating that change to the project infrastructure
will likely run to the end of the calendar year.

For a full list of changes made to DRS since v0.2.0, consult the changelogs
for each version on the [Releases](https://github.com/ga4gh/data-repository-service-schemas/releases)
page.