feat(jsonb): Remove all of sqlite from the tests
Re-arranged how things are set up and fixtures are imported/created, to
avoid re-creating of database tables. When a database driver is
instantiated it automagically goes through the process of creating
tables and running through all the migration steps, because the schema
version is set to 1 on a new database.