DEV-809: fix issue with macOS use of spawn as default (#61)
DEV-809: remove use of multiprocessing due to issues with running python3.8 in macOS

Replace with Thread from threading module. Note that the thread is now run as a daemon with no explicit way to stop. so it stops once the program stops. This works well for tests since the particular fixture that is affected is already a session scoped fixture. It will fail if we ever change to a lower scope